% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack
Implements the knapsack problem using brute force. Also, given parallel = TRUE it implements the brute
force knapsack using parallel computing.}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data frame}

\item{W}{a positive number}

\item{parallel}{initialized to FALSE, if TRUE uses parallel computing}
}
\value{
a list
}
\description{
Brute force knapsack
Implements the knapsack problem using brute force. Also, given parallel = TRUE it implements the brute
force knapsack using parallel computing.
}
\examples{
x <- data.frame(v = c(3, 4, 10), w = c(10, 12, 7))
brute_force_knapsack(x, 10)
}
