% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_v2.R
\name{brute_force_knapsack_v2}
\alias{brute_force_knapsack_v2}
\title{Version two of the brute force knapsack after suggestion at the seminar.
This version of the function, initializes the parallelization only once.}
\usage{
brute_force_knapsack_v2(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data frame}

\item{W}{a positive number}

\item{parallel}{initialized to FALSE, if TRUE uses parallel computing}
}
\value{
a list
}
\description{
Version two of the brute force knapsack after suggestion at the seminar.
This version of the function, initializes the parallelization only once.
}
\examples{
x <- data.frame(v = c(3, 4, 10), w = c(10, 12, 7))
brute_force_knapsack_v2(x, 10)
}
