% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic knapsack
Implement the greedy heuristic version of the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data frame}

\item{W}{a positive number}
}
\value{
a list
}
\description{
Greedy heuristic knapsack
Implement the greedy heuristic version of the knapsack problem
}
\examples{
x <- data.frame(v = c(3, 4, 10), w = c(10, 12, 7))
greedy_knapsack(x, 10)
}
