% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming knapsack
Implements the dynamic programming version of the knapsack problem.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a data frame}

\item{W}{a positive number}
}
\value{
a list
}
\description{
Dynamic programming knapsack
Implements the dynamic programming version of the knapsack problem.
}
\examples{
x <- data.frame(v = c(3, 4, 10), w = c(10, 12, 7))
knapsack_dynamic(x, 10)
}
