% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack objects
Data frame representing the items available to put in the knapsack}
\format{
\describe{
\item{v}{values for the items}
\item{w}{weights for the items}}
}
\usage{
knapsack_objects
}
\description{
Knapsack objects
Data frame representing the items available to put in the knapsack
}
\keyword{datasets}
