% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unoptimized_brute_force_knapsack.R
\name{unoptimized_brute_force_knapsack}
\alias{unoptimized_brute_force_knapsack}
\title{Unoptimized version of the brute force algorithm for the knapsack problem.}
\usage{
unoptimized_brute_force_knapsack(x, W)
}
\arguments{
\item{x}{a data frame}

\item{W}{a positive number}
}
\value{
a list
}
\description{
Unoptimized version of the brute force algorithm for the knapsack problem.
}
\examples{
x <- data.frame(v = c(3, 4, 10), w = c(10, 12, 7))
unoptimized_brute_force_knapsack(x, 10)
}
